cryptobackend/requirements_server.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
gevent==24.2.1
greenlet==3.0.3
idna==3.7
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==1.26.4
pycparser==2.22
requests==2.32.3
schedule==1.2.2
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
websocket==0.2.1
websocket-client==1.8.0
websockets==12.0
zope.event==5.0
zope.interface==6.4.post2
